Refactor package structure, rename types, fields
