+++ pkg/app/app.go
@@ -1,4 +1,4 @@
-package tabula
+package app
 
 import (
 	"fmt"
@@ -6,6 +6,11 @@ import (
 	"time"
 
 	"github.com/go-gl/gl/v2.1/gl"
+	"github.com/gregjohnson2017/tabula-editor/pkg/comms"
+	"github.com/gregjohnson2017/tabula-editor/pkg/config"
+	"github.com/gregjohnson2017/tabula-editor/pkg/image"
+	"github.com/gregjohnson2017/tabula-editor/pkg/menu"
+	"github.com/gregjohnson2017/tabula-editor/pkg/ui"
 	"github.com/gregjohnson2017/tabula-editor/pkg/util"
 	"github.com/veandco/go-sdl2/gfx"
 	"github.com/veandco/go-sdl2/img"
@@ -15,28 +20,21 @@ import (
 // performance debugging metrics
 var imageTotalNs, bbTotalNs, bTotalNs, mlTotalNs, iterations int64
 
-// Config represents the window configuration for the application
-type Config struct {
-	BottomBarHeight int32
-	ScreenHeight    int32
-	ScreenWidth     int32
-}
-
 // Application holds state for the tabula application
 type Application struct {
-	cfg         *Config
-	comps       []UIComponent
-	currHover   UIComponent
+	cfg         *config.Config
+	comps       []ui.Component
+	currHover   ui.Component
 	framerate   *gfx.FPSmanager
-	lastHover   UIComponent
+	lastHover   ui.Component
 	moved       bool
 	postEvtActs chan func()
 	running     bool
 	win         *sdl.Window
 }
 
-// NewApplication returns a newly instantiated application state struct
-func NewApplication(win *sdl.Window, cfg *Config) *Application {
+// New returns a newly instantiated application state struct
+func New(win *sdl.Window, cfg *config.Config) *Application {
 	var fileName string
 	var err error
 	if len(os.Args) == 2 {
@@ -75,15 +73,15 @@ func NewApplication(win *sdl.Window, cfg *Config) *Application {
 		H: 20,
 	}
 
-	bottomBarComms := make(chan imageComm)
-	toolComms := make(chan ImageTool)
+	bottomBarComms := make(chan comms.Image)
+	toolComms := make(chan image.Tool)
 	actionComms := make(chan func())
 
-	iv, err := NewImageView(imageViewArea, fileName, bottomBarComms, toolComms, cfg)
+	iv, err := image.NewView(imageViewArea, fileName, bottomBarComms, toolComms, cfg)
 	errCheck(err)
 	bottomBar, err := NewBottomBar(bottomBarArea, bottomBarComms, cfg)
 	errCheck(err)
-	openButton, err := NewButton(buttonAreaOpen, cfg, "Open File", func() {
+	openButton, err := menu.NewButton(buttonAreaOpen, cfg, "Open File", func() {
 		// TODO fix spam click crash bug
 		newFileName, err := util.OpenFileDialog(win)
 		if err != nil {
@@ -92,30 +90,26 @@ func NewApplication(win *sdl.Window, cfg *Config) *Application {
 		}
 		go func() {
 			actionComms <- func() {
-				err = iv.loadFromFile(newFileName)
+				err = iv.LoadFromFile(newFileName)
 				errCheck(err)
 			}
 		}()
 	})
-	centerButton, err := NewButton(buttonAreaCenter, cfg, "Center Image", func() {
+	centerButton, err := menu.NewButton(buttonAreaCenter, cfg, "Center Image", func() {
 		go func() {
 			actionComms <- func() {
-				iv.centerImage()
+				iv.CenterImage()
 			}
 		}()
 	})
 	centerButton.SetHighlightBackgroundColor([4]float32{1.0, 0.0, 0.0, 1.0})
 	centerButton.SetDefaultTextColor([4]float32{0.0, 0.0, 1.0, 1.0})
 
-	catMenuList := NewMenuList(cfg, false)
-	toolsMenuList := NewMenuList(cfg, false)
+	catMenuList := menu.NewMenuList(cfg, false)
+	toolsMenuList := menu.NewMenuList(cfg, false)
 
-	menuBar := NewMenuList(cfg, true)
-	menuItems := []struct {
-		str string
-		ml  *MenuList
-		act func()
-	}{
+	menuBar := menu.NewMenuList(cfg, true)
+	menuItems := []menu.Definition{
 		{"cat", catMenuList, func() { fmt.Println("cat") }},
 		{"Tools", toolsMenuList, func() {}},
 	}
@@ -123,47 +117,35 @@ func NewApplication(win *sdl.Window, cfg *Config) *Application {
 		panic(err)
 	}
 
-	kittenMenuList := NewMenuList(cfg, false)
-	catSubmenuItems := []struct {
-		str string
-		ml  *MenuList
-		act func()
-	}{
-		{"kitty", &MenuList{}, func() { fmt.Println("kitty") }},
+	kittenMenuList := menu.NewMenuList(cfg, false)
+	catSubmenuItems := []menu.Definition{
+		{"kitty", &menu.MenuList{}, func() { fmt.Println("kitty") }},
 		{"kitten", kittenMenuList, func() { fmt.Println("kitten") }},
 	}
-	if err = catMenuList.SetChildren(0, menuBar.area.H, catSubmenuItems); err != nil {
+	if err = catMenuList.SetChildren(0, menuBar.GetBoundary().H, catSubmenuItems); err != nil {
 		panic(err)
 	}
 
-	kittenSubmenuItems := []struct {
-		str string
-		ml  *MenuList
-		act func()
-	}{
-		{"Mooney", &MenuList{}, func() { fmt.Println("Mooney") }},
-		{"Buttercup", &MenuList{}, func() { fmt.Println("Buttercup") }},
-		{"Sunny", &MenuList{}, func() { fmt.Println("Sunny") }},
+	kittenSubmenuItems := []menu.Definition{
+		{"Mooney", &menu.MenuList{}, func() { fmt.Println("Mooney") }},
+		{"Buttercup", &menu.MenuList{}, func() { fmt.Println("Buttercup") }},
+		{"Sunny", &menu.MenuList{}, func() { fmt.Println("Sunny") }},
 	}
-	if err = kittenMenuList.SetChildren(catMenuList.area.W, menuBar.area.H+catMenuList.area.H/2, kittenSubmenuItems); err != nil {
+	if err = kittenMenuList.SetChildren(catMenuList.GetBoundary().W, menuBar.GetBoundary().H+catMenuList.GetBoundary().H/2, kittenSubmenuItems); err != nil {
 		panic(err)
 	}
 
-	toolsSubmenuItems := []struct {
-		str string
-		ml  *MenuList
-		act func()
-	}{
-		{"No tool", &MenuList{}, func() {
+	toolsSubmenuItems := []menu.Definition{
+		{"No tool", &menu.MenuList{}, func() {
 			// clear the image view tool
-			go func() { toolComms <- EmptyTool{} }()
+			go func() { toolComms <- image.EmptyTool{} }()
 		}},
-		{"Pixel selection tool", &MenuList{}, func() {
+		{"Pixel selection tool", &menu.MenuList{}, func() {
 			// set the image view tool to the pixel selection tool
-			go func() { toolComms <- PixelSelectionTool{} }()
+			go func() { toolComms <- image.PixelSelectionTool{} }()
 		}},
 	}
-	if err = toolsMenuList.SetChildren(0, menuBar.area.H, toolsSubmenuItems); err != nil {
+	if err = toolsMenuList.SetChildren(0, menuBar.GetBoundary().H, toolsSubmenuItems); err != nil {
 		panic(err)
 	}
 
@@ -175,7 +157,7 @@ func NewApplication(win *sdl.Window, cfg *Config) *Application {
 
 	return &Application{
 		running:     false,
-		comps:       []UIComponent{iv, bottomBar, openButton, centerButton, menuBar},
+		comps:       []ui.Component{iv, bottomBar, openButton, centerButton, menuBar},
 		cfg:         cfg,
 		postEvtActs: actionComms,
 		framerate:   framerate,
@@ -227,13 +209,13 @@ func (app *Application) PostEventActions() {
 		comp.Render()
 		ns := sw.StopGetNano()
 		switch comp.(type) {
-		case *ImageView:
+		case *image.View:
 			imageTotalNs += ns
 		case *BottomBar:
 			bbTotalNs += ns
-		case *Button:
+		case *menu.Button:
 			bTotalNs += ns
-		case *MenuList:
+		case *menu.MenuList:
 			mlTotalNs += ns
 		}
 	}
@@ -331,7 +313,7 @@ func (app *Application) Running() bool {
 func (app *Application) Quit() {
 	avgNs := int64(float64(imageTotalNs) / float64(iterations))
 	avg := time.Duration(int64(time.Nanosecond) * avgNs)
-	fmt.Printf("ImageView avg:\t%v\n", avg)
+	fmt.Printf("image.View avg:\t%v\n", avg)
 	avgNs = int64(float64(bbTotalNs) / float64(iterations))
 	avg = time.Duration(int64(time.Nanosecond) * avgNs)
 	fmt.Printf("BottomBar avg:\t%v\n", avg)
@@ -342,7 +324,7 @@ func (app *Application) Quit() {
 	avg = time.Duration(int64(time.Nanosecond) * avgNs)
 	fmt.Printf("MenuList avg:\t%v\n", avg)
 
-	// free UIComponent SDL assets
+	// free ui.Component assets
 	for _, comp := range app.comps {
 		comp.Destroy()
 	}
@@ -350,13 +332,6 @@ func (app *Application) Quit() {
 	app.win.Destroy()
 	sdl.Quit()
 	img.Quit()
-}
-
-func inBounds(area *sdl.Rect, x int32, y int32) bool {
-	if x < area.X || x >= area.X+area.W || y < area.Y || y >= area.Y+area.H {
-		return false
-	}
-	return true
 }
 
 func errCheck(err error) {
