+++ pkg/comms/image.go
@@ -1,0 +1,9 @@
+package comms
+
+import "github.com/veandco/go-sdl2/sdl"
+
+type Image struct {
+	FileName string
+	MousePix sdl.Point
+	Mult     float64
+}
