+++ pkg/menu/list.go
@@ -1,141 +1,25 @@
-package tabula
+package menu
 
 import (
 	"fmt"
 	"math"
 
+	"github.com/gregjohnson2017/tabula-editor/pkg/config"
+	"github.com/gregjohnson2017/tabula-editor/pkg/font"
 	"github.com/veandco/go-sdl2/sdl"
 )
-
-var _ UIComponent = UIComponent(&MenuList{})
-
-// MenuEntry is the clickable entry which opens a MenuList
-type MenuEntry struct {
-	enabled bool
-	button  *MenuBarButton
-	list    *MenuList
-	action  func()
-}
-
-// MenuBarButton defines an interactive button, but redefines OnClick to perform action on press, not release
-type MenuBarButton struct {
-	*Button
-}
-
-// OnClick is called when the user clicks within the UIComponent's region
-func (mbb *MenuBarButton) OnClick(evt *sdl.MouseButtonEvent) bool {
-	if evt.Button == sdl.BUTTON_LEFT && evt.State == sdl.PRESSED {
-		mbb.pressed = true
-		mbb.action()
-	} else if evt.Button == sdl.BUTTON_LEFT && evt.State == sdl.RELEASED && mbb.pressed {
-		mbb.pressed = false
-		// TODO add release action
-	}
-	return true
-}
-
-// NewMenuEntry returns the struct with the given label and list
-func NewMenuEntry(area *sdl.Rect, label string, list *MenuList, cfg *Config, act func()) (MenuEntry, error) {
-	if cfg == nil {
-		return MenuEntry{}, fmt.Errorf("NewMenuEntry found nil Config")
-	}
-	if list == nil && act == nil {
-		return MenuEntry{}, fmt.Errorf("NewMenuEntry needs a list and/or an action")
-	}
-	var btn *Button
-	var err error
-	if btn, err = NewButton(area, cfg, label, act); err != nil {
-		return MenuEntry{}, err
-	}
-	return MenuEntry{
-		enabled: false,
-		button:  &MenuBarButton{btn},
-		list:    list,
-	}, nil
-}
-
-// Destroy calls destroy on underlying UIComponents
-func (me MenuEntry) Destroy() {
-	me.button.Destroy()
-	me.list.Destroy()
-}
-
-// InBoundary returns whether a point is in this UIComponent's bounds
-func (me MenuEntry) InBoundary(pt sdl.Point) bool {
-	if me.button.InBoundary(pt) {
-		return true
-	}
-	if me.enabled && me.list != nil && me.list.InBoundary(pt) {
-		return true
-	}
-	return false
-}
-
-// GetBoundary returns the underlying button's boundary
-func (me MenuEntry) GetBoundary() *sdl.Rect {
-	return me.button.GetBoundary()
-}
-
-// OnEnter calls the underlying button's OnEnter method
-func (me *MenuEntry) OnEnter() {
-	me.button.OnEnter()
-}
-
-// OnLeave calls the underlying button's OnLeave method
-func (me *MenuEntry) OnLeave() {
-	me.button.OnLeave()
-	me.list.OnLeave()
-}
-
-// OnClick calls the underlying button's OnClick method
-func (me *MenuEntry) OnClick(evt *sdl.MouseButtonEvent) bool {
-	if evt.Button != sdl.BUTTON_LEFT || evt.State != sdl.PRESSED {
-		return true
-	}
-	if inBounds(me.GetBoundary(), evt.X, evt.Y) {
-		me.button.OnClick(evt)
-		me.enabled = !me.enabled
-		return true
-	}
-	if e, err := me.list.GetEntryAt(evt.X, evt.Y); err == nil {
-		e.OnClick(evt)
-	}
-	return true
-}
-
-// Render calls the underlying button's render function
-func (me MenuEntry) Render() {
-	me.button.Render()
-	if me.list != nil && me.enabled {
-		me.list.Render()
-	}
-}
-
-// OnResize calls the underlying UIComponents' OnResize function
-func (me MenuEntry) OnResize(x, y int32) {
-	me.button.OnResize(x, y)
-	me.list.OnResize(x, y)
-}
-
-// OnMotion is called when the cursor moves within the UIComponent's region - bad comment
-func (me *MenuEntry) OnMotion(evt *sdl.MouseMotionEvent) bool {
-	if me.list.InBoundary(sdl.Point{X: evt.X, Y: evt.Y}) {
-		me.list.OnMotion(evt)
-	}
-	return true
-}
 
 // MenuList is the horizontal menu bar
 type MenuList struct {
 	area    *sdl.Rect
-	cfg     *Config
+	cfg     *config.Config
 	entries []MenuEntry
 	hover   *MenuEntry
 	horiz   bool
 }
 
 // NewMenuList returns a pointer to a new MenuList struct that implements UIComponent
-func NewMenuList(cfg *Config, horiz bool) *MenuList {
+func NewMenuList(cfg *config.Config, horiz bool) *MenuList {
 	return &MenuList{
 		area:  &sdl.Rect{},
 		cfg:   cfg,
@@ -143,17 +27,19 @@ func NewMenuList(cfg *Config, horiz bool) *MenuList {
 	}
 }
 
+type Definition struct {
+	Str string
+	Ml  *MenuList
+	Act func()
+}
+
 // SetChildren registers a set of menu entries with the menu bar
-func (ml *MenuList) SetChildren(offx int32, offy int32, childs []struct {
-	str string
-	ml  *MenuList
-	act func()
-}) error {
+func (ml *MenuList) SetChildren(offx int32, offy int32, childs []Definition) error {
 	for _, e := range ml.entries {
 		e.Destroy()
 	}
 	ml.entries = make([]MenuEntry, 0, len(childs))
-	font, err := loadFontTexture("NotoMono-Regular.ttf", 14)
+	fnt, err := font.LoadFontTexture("NotoMono-Regular.ttf", 14)
 	if err != nil {
 		return err
 	}
@@ -162,7 +48,7 @@ func (ml *MenuList) SetChildren(offx int32, offy int32, childs []struct {
 	// normalize height or width
 	var max int32
 	for _, c := range childs {
-		w, h := calcStringDims(c.str, font)
+		w, h := font.CalcStringDims(c.Str, fnt)
 		w32 := int32(math.Ceil(w)) + 14
 		h32 := int32(math.Ceil(h)) + 10
 		if ml.horiz {
@@ -178,7 +64,7 @@ func (ml *MenuList) SetChildren(offx int32, offy int32, childs []struct {
 	// populate list of menu entries with appropriate boundaries
 	var off int32
 	for _, child := range childs {
-		w, h := calcStringDims(child.str, font)
+		w, h := font.CalcStringDims(child.Str, fnt)
 		w32 := int32(math.Ceil(w)) + 14
 		h32 := int32(math.Ceil(h)) + 10
 		var area *sdl.Rect
@@ -187,7 +73,7 @@ func (ml *MenuList) SetChildren(offx int32, offy int32, childs []struct {
 		} else {
 			area = &sdl.Rect{X: ml.area.X, Y: ml.area.Y + off, W: max, H: h32}
 		}
-		entry, err := NewMenuEntry(area, child.str, child.ml, ml.cfg, child.act)
+		entry, err := NewMenuEntry(area, child.Str, child.Ml, ml.cfg, child.Act)
 		if err != nil {
 			return err
 		}
